RNG-91 Implemented compute shader pipelines and equirect env map loading with texture loader

+++ Src/Core/Src/Blob.cpp
@@ -2,7 +2,7 @@
 
 namespace core
 {
-    Blob::Blob(const void* data, uint32_t size)
+    Blob::Blob(const void* data, size_t size)
     {
         m_data.resize(size);
         std::memcpy(m_data.data(), data, size);
